Update docker instructions with timezone
Putting a note on how to get the TZ=Your/Timezone info here, both how to get the current one on the system (eg. local) and how to get a list (eg. portainer).

I have no affiliation with the one I linked (https://timezonedb.com/time-zones) but it looked easier to use than the one on wikipedia and seemed committed to being free-to-use.

Also fixed up whitespace to match between the sections.